move hybrid head to head.py
